+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Lower the minimum version requirements in [NetVips.Extensions](https://www.nuget.org/packages/NetVips.Extensions/) package:
   - `System.Drawing.Common` to 8.0.0 (only required when targeting .NET 6 or later).
   - .NET Framework version to 4.6 (`net46`).
+- Deprecate AOT-incompatible `Image.NewFromMemory(Array, int, int, int, Enums.BandFormat)` overload.
 
 ### Fixed
 - Fix size calculation in `Image.NewFromMemory()` for non-byte element types ([#264](https://github.com/kleisauke/net-vips/issues/264)).